refactor(test): extract JSON-specific tests to separate file
This commit moves all JSON-related tests from response_test.go to a new file
response_json_test.go to improve test organization and readability.

Benefits:
- Separates concerns between generic response tests and JSON-specific tests
- Improves maintainability by grouping similar tests together
- Enhances test suite readability
